fix: restart_server.sh 스크립트 수정
- 현재 실행중인 애플리케이션이 없는 경우, 발생하는 스크립트 에러 fix

+++ scripts/restart_server.sh
@@ -3,6 +3,14 @@ echo 'Running restart_server'
 echo '======================'
 
 service nginx restart
+echo "> 현재 구동중인 애플리케이션 pid 확인"
 IDLE_PID=$(pgrep -f around)
-kill -15 "$IDLE_PID"
-sleep 5
+
+if [ -z "$IDLE_PID" ]
+then
+  echo "> 현재 구동중인 애플리케이션이 없으므로 종료하지 않습니다."
+else
+  echo "> kill -15 $IDLE_PID"
+  kill -15 "$IDLE_PID"
+  sleep 5
+fi
